refactor(ynu): code resuse for readability
