+++ lib/routes/universities/ynu/grs/qttz.js
@@ -1,6 +1,6 @@
 const got = require('@/utils/got');
 const cheerio = require('cheerio');
-const url = require('url');
+const utils = require('../utils');
 
 // 云南大学研究生院其他通知
 module.exports = async (ctx) => {
@@ -25,45 +25,7 @@ module.exports = async (ctx) => {
             author: dep[category - 1],
         }))
         .get();
-
-    const out = await Promise.all(
-        list.map(async (item) => {
-            const { path, title, author } = item;
-            const link = path.indexOf('http://') === -1 ? host + path : path;
-            const cache = await ctx.cache.get(link);
-            if (cache) {
-                return Promise.resolve(JSON.parse(cache));
-            }
-
-            const response = await got({
-                method: 'get',
-                url: link,
-                headers: {
-                    Referer: host,
-                },
-            });
-
-            const $ = cheerio.load(response.data);
-            const description =
-                $('.v_news_content').html() &&
-                $('.v_news_content')
-                    .html()
-                    .replace(/src="\//g, `src="${url.resolve(host, '.')}`)
-                    .replace(/href="\//g, `href="${url.resolve(host, '.')}`)
-                    .trim();
-
-            const single = {
-                pubDate: new Date($('#times').text().slice(5)).toUTCString(),
-                author: author,
-                link: link,
-                title: title,
-                description: description,
-            };
-            ctx.cache.set(link, JSON.stringify(single));
-            return Promise.resolve(single);
-        })
-    );
-
+    const out = await utils.processPages({ list: list, caches: ctx.cache, host: host, department: 'grs' });
     ctx.state.data = {
         title: '云南大学研究生院' + dep[category - 1] + '通知',
         link: host,
