+++ lib/routes/universities/ynu/jwc/zytz.js
@@ -1,6 +1,6 @@
 const got = require('@/utils/got');
 const cheerio = require('cheerio');
-const url = require('url');
+const utils = require('../utils');
 
 // 云南大学教务处主要通知
 module.exports = async (ctx) => {
@@ -31,52 +31,7 @@ module.exports = async (ctx) => {
         }))
         .get();
 
-    const out = await Promise.all(
-        list.map(async (item) => {
-            const { path, title, author } = item;
-            const link = path.indexOf('http://') === -1 ? host + path : path;
-            const cache = await ctx.cache.get(link);
-            if (cache) {
-                return Promise.resolve(JSON.parse(cache));
-            }
-
-            const response = await got({
-                method: 'get',
-                url: link,
-                headers: {
-                    Referer: host,
-                },
-            });
-
-            const $ = cheerio.load(response.data);
-            let description =
-                $('.v_news_content').html() &&
-                $('.v_news_content')
-                    .html()
-                    .replace(/src="\//g, `src="${url.resolve(host, '.')}`)
-                    .replace(/href="\//g, `href="${url.resolve(host, '.')}`)
-                    .trim();
-            if ($('#vsb_content').siblings().has('ul li').length > 0) {
-                const attachment = $('#vsb_content')
-                    .siblings('ul')
-                    .html()
-                    .replace(/src="\//g, `src="${url.resolve(host, '.')}`)
-                    .replace(/href="\//g, `href="${url.resolve(host, '.')}`)
-                    .trim();
-                description = description.concat('\n' + attachment);
-            }
-            const single = {
-                pubDate: new Date($('.taitshj').text().slice(5, 15)).toUTCString(),
-                author: author,
-                link: link,
-                title: title,
-                description: description,
-            };
-            ctx.cache.set(link, JSON.stringify(single));
-            return Promise.resolve(single);
-        })
-    );
-
+    const out = await utils.processPages({ list: list, caches: ctx.cache, host: host, department: 'jwc' });
     ctx.state.data = {
         title: '云南大学教务处' + dep[category - 1] + '通知',
         link: host,
